[trace_processing] Add total wall time metric
Some metrics such as CPU load average make the most sense when the total
time taken by test runs is the same. Additionally, we have sometimes
encountered issues where one of the symptoms is the system stalling
during a test case, often in a flaky manner. This change adds the total
duration of a trace as a metric so that we can surface that information
more readily.

Bug: b/161845125
Change-Id: I38ea74621780087e9589a5f58b181e391763660e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409281
Reviewed-by: Nathan Rogers <nathanrogers@google.com>
Testability-Review: Nathan Rogers <nathanrogers@google.com>
Commit-Queue: Brian Hamrick <bhamrick@google.com>