[camera][sensor] Add analog gain support
This CL adds support for programming the analog gain in the sensor.
To promote reuse of the code, details of the analog gain encoding
and limits are read from standard sensor registers.

As this is the first piece of exposure setting implementation,
it also introduces the deferral of register writes and their
synchronization through the grouped parameter hold mechanism.

In support of these additions, a 16-bit I2C write routine was added.

Note that this CL only adds support to the CameraSensor2 interface.
This is the interface currently used by the ISP, and the older
interface will soon be removed.

Change-Id: Ib786f6ff21de6038423694f1f59331c6335748ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412554
Commit-Queue: John Sasinowski <jsasinowski@google.com>
Reviewed-by: Nic Zoghb <nzo@google.com>
Testability-Review: Nic Zoghb <nzo@google.com>