[camera][sensor] Add integration time support
This CL adds support for programming the integration time in the sensor.

As part of this effort, the calculation of lines per second was
refactored into a separate method. To facilitate this,
GetRegisterValueFromSequence() also became a method, and both moved to
imx227.cc.

This CL also tidies up more of the constants, moving them to constants.h
and eliminating duplicates.

Testing:
    Tested visually.
    fxb/50737 tracks the development of a new sensor unit test framework.

Change-Id: Ie2a3e58612379c6193f25de2f13616b0fc7bbe4e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412402
Commit-Queue: John Sasinowski <jsasinowski@google.com>
Reviewed-by: Nic Zoghb <nzo@google.com>
Testability-Review: Nic Zoghb <nzo@google.com>