[amlogic_decoder] Keep cores powered up for entire lifetime of decoders
The current code powers off a core when an instance is swapped out and
powers it on when the instance is swapped in. This causes a lot of short
power-off/power-on transitions, which seems to cause some rare hangs. A
workaround added to increase the time of those transitions slows down
decoding, so we should reduce those transitions as much as possible.

In practice, idle decoders remain scheduled and powered on, so the only
time this change will cause a core to stay on when it would otherwise be
powered off (for long periods of time) is when context switching between
VP9 and H264, where before the unused core would be powered off.

Bug: 57380

Change-Id: I08f5f258717e2a82d0903a092f1f941a7dab75aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412400
Reviewed-by: Dustin Green <dustingreen@google.com>
Testability-Review: John Bauman <jbauman@google.com>
Commit-Queue: John Bauman <jbauman@google.com>