[camera] Fix new flake in Device test.
This change fixes a flake introduced in fxr/411922 related to stuck-
client buffer recycling. Previously, the test did not wait for a frame
to be received before sending subsequent frames. This caused the test to
erroneously pass most of the time, because it only verified that a
specific frame was eventually recycled, not recycled in a specific
order. Additionally, the test was using inconsistent values for min-
camping-buffers, which could lead to deadlock. This change fixes the
issue by waiting for each frame to be correctly received or recycled,
and by using a common constant value for min buffer count.

Multiply: fuchsia-pkg://fuchsia.com/camera_tests#meta/device_test.cmx: 100
Fixed: 57256
Test: Ran DeviceTest.GetFrames with repeat=10000.
Change-Id: Ifac6c79a2e01b8fe8215137cd90a73e7bc08ba8c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413022
Reviewed-by: Nic Zoghb <nzo@google.com>
Testability-Review: Nic Zoghb <nzo@google.com>
Commit-Queue: Matt Sandy <msandy@google.com>