[zbitl] Fix and revamp CRC32 checking
The CRC-checking mode was completely broken.  It computed the
payload CRC regardless of whether the item header had the CRC flag
set.  It then never checked the CRC it computed against the header.
This went unnoticed due to the confluence of two issues.  First, the
test is permissive (by design), assuming any non-storage error is
the expected CRC mismatch error.  Second, the test data in the
corpus was bad, having a corrupted header instead of a corrupted
payload.  The test now checks for the exact error string, and the
test data is now correct.

The StorageTraits::Crc32 method is replaced with a generalized Read
method that takes a callback function.  This will be used to support
other library-based payload access like ZBI_FLAG_STORAGE_COMPRESSED
decompression in the future.

Bug: 49438
Change-Id: Ied652c2d3dbc881dc617d2b58bb7acbbfbf013cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412774
Commit-Queue: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Testability-Review: Joshua Seaton <joshuaseaton@google.com>