[GN SDK] Using architecture specific host tools
This changes the references to host tools in the build rule templates
and scripts to use the architecture specific tools.

If the host_cpu variable in the build is not supported, the x64
version is used.

For the scripts in bin/ the architecture is determined by mapping
`uname -m` to x64 or arm64. If the mapping fails, the tool path used
is tools/$(uname -m).

Bug: 56744

Change-Id: I9d55447cb6c4aee44975a9f765c8bac64cb79e59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412474
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Testability-Review: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Wayne Piekarski <waynepie@google.com>