[wlan][sl4f] Fix locking around hanging get_update
Previously if a get_update hung from never getting an update from the
policy layer, the lock around the inner facade would still be held by
the hanging get_update even if the consumer timed out from the update.
After that, no other commands would work because they would be waiting
for the lock. Now the update listener is behind a separate lock and the
lock is behind a reference so that the lock and update stream can be
reset for each test.

Bug: 56323
Test: act.py -c <config> -tc SavedNetworksTest
	tested that only the one test fails if it never gets an update
Change-Id: I8680d0802a1186f683689ecd1f02ecfb29800b4e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407639
Commit-Queue: Naomi Mccracken <nmccracken@google.com>
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Reviewed-by: Joe Brennan <jmbrenna@google.com>
Testability-Review: Kevin Sakuma <sakuma@google.com>