[acpi][x64] Remove AcpiWalker class.
Remove the AcpiWalker class and replace it with a call to a new
utility function which allows us to walk the ACPI namespace using a
lambda with capture storage for state instead of putting the state
into a class.

Also, get the PCI code out of the business of needing to care about
the internals of the AcpiWalker class.

Change-Id: I28e3c0b2985fb524c5f480365457cd648776b2e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411940
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>