[brcmfmac] Use firmware provided reason code
In 57315, when firmware reported a DEAUTH_IND event, it provided reason
code 6 and 7. However, the reason code reported to wlanstack is 3. Turns
out it was created based on the event type.

The correct behavior seems to be forwarding the reason code directly
as-is.

Bug: 57315
Test: Included

Change-Id: I4b820c8f21d61b24b5f72e068c438ee6a9e8112b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412621
Commit-Queue: Eric Wang <eyw@google.com>
Testability-Review: Karthik Krishnaswami <karthikrish@google.com>
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>