[system-updater] Generate a stream of states
In preparation to yield status updates over FIDL when installing an
update, this change modifies the task that installs the update to yield
that data as it installs the update using a a State enum that closely
matches the fuchsia.update.installer.State union.

This change also modifies the state machine defined in the FIDL to
specify that the system updater always progresses through the
WaitToReboot state before entering the Reboot state.

As this change modifies the on-disk format of the update history, the
prior data will be lost when upgrading past this CL.  Otherwise, no
functional changes are expected.

Bug: 55407
Change-Id: I0893769a07c109801a47270715940e95cf8c3f6e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410913
Testability-Review: Kevin Wells <kevinwells@google.com>
Commit-Queue: Kevin Wells <kevinwells@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
API-Review: Dan Johnson <computerdruid@google.com>