[toolchain] Pass along library path for libLLVM
Without this change we are unable to build wasm32-unknown-unknown shared
libraries using our toolchain. For some reason that combination of
settings causes us to look for libLLVM-....so which is not in the path
for some reason.

This workaround unblocks us for now.

This CL also adds //src/diagnostics/lib/triage/wasm:wasm to
//src/diagnostics:tests to make sure wasm can build.

Bug: 56076

Change-Id: I9dfcc761c741fa51a79ea1852cb73bcfb218302d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406702
Commit-Queue: Christopher Johnson <crjohns@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>