[fidl] Change "resource" keyword to "resource_definition"
The "resource" keyword will be used for FTP-057 as a modifier on
structs, tables, and unions. But it is already used for defining new
kinds of resources, i.e. handle generalization. It is difficult to
overload the keyword for both usages because the parser is designed for
only one lookahead, so the easiest solution is to use different
keywords. Since the FTP-057 usage will be far more common than declaring
new kinds of resources, we prefer to reserve "resource" for the former,
and use the longer keyword "resource_definition" for the latter.

Test: fx test fidl-compiler --host
Change-Id: Ic12305c1d2407fe922901c4813b552c95413f196
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412734
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Scott Graham <scottmg@google.com>
API-Review: Pascal Perez <pascallouis@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>