[fidl][llcpp] refactor async requests
Use OwnedRequest to send the asynchronous request. That means that, now,
request <= 512 bytes are allocated on the stack.

Refactor dispatch. Now the handling of the contexts is generic. That
means that we no longer have a switch for the replies' ordinal (we only
have a switch for events).

A side effect is that fixes a vulnerability which could have been
exploted by a malicious server. With the previous implementation, a
reply was decoded using the ordinal within the message header then, the
result was used (with a cast) using the txid. An attacker could have
send a reply with an ordinal different from the request. This could have
crash the client or open a back door.

The asynchronous version is barely used. We only spare 1240 bytes.

Change-Id: I94c9d976593ca382158f8255bf860d2641a595fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411924
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Madhav Iyengar <madhaviyengar@google.com>
Commit-Queue: Vincent Belliard <vbelliard@google.com>
Testability-Review: Yifei Teng <yifeit@google.com>