[fdr] Show an overlay while counting down to reset
Show the usual overlay when counting down to a factory data reset. Use
F11 as volume down and F12 as volume up to test on a device without
consumer control keys.

Also added copy to the keyboard phase as there’s no reason for it not
to be copyable.

This is unlikely to be the last UI change to this program, but it is
worth committing as is.

Testing: manual
Change-Id: I27e7f367158f92b704269ea820668d1a6698de7c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412126
Reviewed-by: Dragos Tiselice <dtiselice@google.com>
Reviewed-by: Mike Comfoltey <comfoltey@google.com>
Testability-Review: Dragos Tiselice <dtiselice@google.com>
Commit-Queue: Rob Tsuk <robtsuk@google.com>