[dart] fix crash on malformed pubspec
The package importer is crashing because of a file with tabs
at the end of the line. This is not allowed in pubspec files
but we cannot skip the packages because they are required by
flutter. We trim the trailing tab and continue.

Change-Id: I8f09310ff2311efef35fedb0a8e067872168d27c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413600
Reviewed-by: Darren Chan <chandarren@google.com>
Testability-Review: Chase Latta <chaselatta@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>