[FDF] Exit gracefully when stop event completes
In component lifecycle stop event handler, driver manager suspends all drivers and
If the termination state is Reboot/shutdown, it also issues reboot/shutdown syscall.
At the end of the handler, closing the lifecycle handle causes the component manager
to kill driver manager. When driver manager which is marked critical is killed, the
root job watcher in kernel issues a reboot. If mexec is switched to use orderly shutdown,
this results in an undesired behavior where we want to mexec the system, not reboot.

This changeset exits the driver manager gracefully instead of closing
the lifecycle event handle. If we do so, kernel will not reboot the
system.

Test: fx reboot. fx netboot after moving mexec lib to orderly shutdown.

Change-Id: I5680d842f483bec1388e6bbc27e0153982139af3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412222
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>
Commit-Queue: Ruchira Ravoori <ravoorir@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>