[time] Fix cobalt logging bug.
Although the first event was logging fine the CobaltMetrics object and
the Task it contained was being dropped as soon as time synchronization
succeeded, meaning the sucess event was never actually sent.

Longer term time sync will be a loop and so we won't have this problem,
but for now the easiest solution is just detach the task.

Bug: 57677
Change-Id: I5f52658e2394b2b1866438300941a7ff453dd3f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413596
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Testability-Review: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>