[aml][thermal] Divide S905d2g Thermal into New and Legacy
aml-thermal-s905d2g is currently used for three different SoCs:
    + s905d2
    + t931
    + s905d3

This driver is currently responsible for the following activities:
    + Monitoring Thermals
    + Programming Power Domains
    + Programming Clock Domains
    + Executing DVFS

Ideally we would like for this driver to only be responsible for
monitoring temperature sensors and allow the power, clock, and cpu
drivers to perform the other three actions respectively.

The s905d2 variant of the driver is already ready to have the
extraneous functionality stripped whereas the 905d3 and t931 variants
still rely on that functionality to exist. In order to move forward
with removing the extraneous functionality, this patch bifurcates the
aml-thermal driver into the legacy variant for 905d3 and t931 and a
modern version for 905d2.

Driver was tested by booting on a Sherlock and Astro device and
verifying that the appropriate was bound.

Further verification was done by ensuring that the following test suites
still passed:
    + aml-thermal-s905d2g-legacy-test
    + aml-thermal-s905d2g-test

Change-Id: I20b34354c8c7c0abe1d9a5d822a5e1ec9ccd4b60
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401379
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>
Reviewed-by: Patrick Shickel <pshickel@google.com>
Commit-Queue: Gurjant Kalsi <gkalsi@google.com>
Testability-Review: Braden Kell <bradenkell@google.com>