[fuchsia-component] fail nested env creation with sub dirs
Nested environments are provided services from the root of a ServiceFs.
Developers are used to putting services inside of subdirectories. If a
developer creates a nested environment but puts the services for it in a
subdirectory, the nested environment will silently not receive these
services.

Update the create_nested_environment functions to check for sub
directories, and produce an error if they exist.

Change-Id: Ibe56a169e7150bab6e2c4af047e009cdf5ebb58e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412655
Reviewed-by: Bryce Lee <brycelee@google.com>
Reviewed-by: William Xiao <wxyz@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Derek Gonyeo <dgonyeo@google.com>
Testability-Review: Gary Bressler <geb@google.com>