[runtests] Remove --restrict-logs flag.
We are removing --restrict-logs flag from run-test-component, so need to
remove it from runtests. There is a alternative of --restrict-logs we
are developing but runtests will not support it as it is only used on
profile builders and there is no value added to fail the tests using
this/alternate flag on profile builders for now.

BUG=49735

Change-Id: I72ae7880649d09dcb9108b4bb59889ab431a2be0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411834
Commit-Queue: Ankur Mittal <anmittal@google.com>
Reviewed-by: Gary Miguel <garymm@google.com>
Testability-Review: Gary Miguel <garymm@google.com>