[SetUI] Store client/request with active requests
When sending a request through the registry, the original request
and the client should be stored so the events can be retried in case
of a failure (retries not in this CL).

This also fixes a bug in the regeneration test. A new handler is
created so that the event loop for the new request runs in the registry.

Add a note about how to prevent leaks when the retry behavior is
implemented.

Tests: fx test -o setui_service_tests
Bug: 57168
Fixed: 57166

Change-Id: I7709203ecbf99c23225b25000e0aff2bc081c11e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413338
Commit-Queue: Paul Faria <paulfaria@google.com>
Reviewed-by: Bryce Lee <brycelee@google.com>
Testability-Review: Bryce Lee <brycelee@google.com>