[camera][bin] Connect FactoryServer to fuchsia.factory.camera.Isp
This CL changes the FactoryServer to pass incoming calls that target
the ISP or sensor to the corresponding calls on the ISP interface.

To necessitate this, the ISP's protodef was edited, and the device
was added to the component manifest.

A change was also made to the Controller protocol to simply pass a
channel to the ISP protocol instead of duplicating code.

This change also includes a fix for a linkage bug in the factory
server and an i2c write operation in the imx227 driver.

Test: visually confirmed that calls are relayed to the ISP by toggling
the sensor test patterns. This was done via an offline CL that attaches
a CLI to the camera-factory component.

Change-Id: I10a190ee1c09db6ae493687d759e3da5d99916fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412745
Commit-Queue: Nic Zoghb <nzo@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: John Sasinowski <jsasinowski@google.com>
Testability-Review: Matt Sandy <msandy@google.com>