[blackout] add ssh authentication configuration options
move off of fx ssh, and instead use regular ssh. in order for this to
work, we have to configure the ssh client ourselves in a sane way. this
adds a bunch of general configuration options to the ssh command, as
well as adding auth configuration for various situations.

Fixed: 46959
Test: fx test //tools/blackout/blackout-host
Change-Id: I002be96418e17f1a30e9afa5e2cf17176e52d79b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/367433
Reviewed-by: Manali Bhutiyani <manalib@google.com>
Testability-Review: Manali Bhutiyani <manalib@google.com>
Commit-Queue: Nathan Mulcahey <nmulcahey@google.com>