[Settings] Migrate back to watch
This a step in the process to rename Watch2 to Watch. A previous api
change required us to add a version of Watch that didn't return a
result/error, so it had to be renamed Watch2 so that both could
exist for a soft-transition. This change is to migrate back to the name
Watch.

Bug: 55719, 55685
Test: fx test setui_service_tests cobalt_app_unittests

Change-Id: I3248f84970a461501e3ca726db781f6ef6f26347
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414001
Reviewed-by: Cameron Dale <camrdale@google.com>
Testability-Review: Cameron Dale <camrdale@google.com>
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>