[fidlc] Bound conditions on unknown enum member
When a flexible enum member is specified to be some value, allow the use
of the otherwise reserved max value.

Recording the unknown value of a flexible enum to later expose it in the
JSON IR. However, we first need to enable this feature in the compiler.

Test: added
Change-Id: Id15bc3968410730180582672c6d50ac5c51aa6da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410875
Reviewed-by: Mitchell Kember <mkember@google.com>
Testability-Review: Mitchell Kember <mkember@google.com>
Commit-Queue: Pascal Perez <pascallouis@google.com>