[paver] add support for new partition scheme
Provides a standardized partition scheme that devices can use going
forward. Not all devices will be required to use this scheme, but for
cases that we have full control of the partition layout, we'll try to
keep as close to this as possible.

Additionally, this new scheme corrects two slight oddities with the old
strategy:
 * UUIDs were not generated randomly - several runs of bits were
   common between partitions
 * Partitions were uniquely identified by type GUID, whereas in the new
   scheme slotted partitions (e.g. zircon_{a,b,c}) will share the same
   type and be distinguished by name and unique GUID instead

Lastly, the debug names of partitions printed to the console have been
generalized instead of always using the legacy names. This fixes the
harmless but somewhat confusing cases like e.g. the Astro paver logging
that it's updating the "fuchsia-esp" partition which doesn't exist.

Bug: 51200, 51442, 56775
Test: Sherlock pave
Change-Id: I6acd7b383d8386310e29144853d86d451b0d83d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409721
Reviewed-by: Manali Bhutiyani <manalib@google.com>
Reviewed-by: Carl Norum <cjn@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>