[cml] Storage backing dir accepts name or path
Add a `backing_dir` field to the CML `storage` capability that accepts
the name of a directory capability. This will eventually replace `path`
which identifies the capability by `path`.

Putting this in a new field, instead of overloading `path`, simplifies
the implementation of cmc a bit. Also, it's closer to the final syntax
we'll end up with.

Note that the CM and CM-field still uses `storage_path` for both. In
this case, it turned out to be easier to reuse `storage_path` than to
add another field.

Bug: 56604
Change-Id: I5490be3270ca966c2ef2aaa34d9da0d9dbdef733
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413138
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Testability-Review: Derek Gonyeo <dgonyeo@google.com>