[net-cli] Add dhcp start/stop
This change adds net-cli interfaces to start/stop a dhcp client on the
specified interface. This is part of an effort to bring interfaces that
currently only exist in ifconfig into net-cli, with the aim of
eventually deprecating and removing the former.

Fixed: 56759

Change-Id: I78a1848e645f686313889f47526f282497dd4ac8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414356
Commit-Queue: Amanda Tait <atait@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Bert Muthalaly <stijlist@google.com>
Testability-Review: Tamir Duberstein <tamird@google.com>