[GN SDK] Adding absolute root path flag to cipd rule
This adds a flag to the cipd package rule template which causes the
root property of the cipd package definition to use an absolute path vs.
relative to root_build_dir.

This allows the cipd definition to be run outside the GN/ninja
environment, for example in a recipe.

Change-Id: If24eb5789774d262608e20a17fce2c2c08edae65
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414388
Testability-Review: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Jonathan Koren <jdkoren@google.com>