[fidlcat] fix summary
Displays all the sessions found for a given handle. This case is mostly
found within tests when a handle is sent to the same process.

Fixes SyscallEvent::GetValue to be able to also get outline fields.

Fixes display of message summary when the ordinal is not known.

Change-Id: Ide3c2df00f9fafbcb2f6f97bf3c030052a2a5356
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412198
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Testability-Review: Jeremy Manson <jeremymanson@google.com>
Commit-Queue: Vincent Belliard <vbelliard@google.com>