[gn][rust] Massage source paths for verification
When a relative path to a generated source file is used in a Rust build
(include_str!), the depfile emitted by the compiler is relative to the
build directory but not necessarily canonical. For example,
'../../out/default/foo/bar'. This change makes sure that the paths
are massaged to something like 'foo/bar' (given that the build directory
is 'out/default').

Change-Id: I604871d57863d463219b57aea9ce0f8cc47e4613
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414457
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Testability-Review: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Adam Lesinski <adamlesinski@google.com>