[fidl] Combine visitor implementation for fidl_decode and fidl_validate
These are nearly identical, except for a few small details such as:
- Decoder mutates and therefore needs non-const pointers
- The validator only keeps track of handle count, not actual handles

As a result, the implementations can be combined.

The motivation for this is to reduce work when making changes to the
visitors - currently you need to do essentially identical changes
in both of these files.

Test: existing tests for these functions

Change-Id: Icab321b7bdcc62c3c6da514bafca4ac9944f0060
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412456
Reviewed-by: Pascal Perez <pascallouis@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>