[fbl] Ignore that GPArena HeadNode operations aren't lock-free on riscv64
RISC-V does not have a 128 bit compare and swap
https://lists.riscv.org/g/tech/topic/requirements_for_128_bit/76126473?p=,,,20,0,0,0::recentpostdate%2Fsticky,,,20,2,20,76126473

In https://fuchsia-review.googlesource.com/c/fuchsia/+/306989/ the
GPArena was refactored to use a 128 bit atomic operation with a static
check that this is done lock-free, using a compare and swap. However,
being lock-free is only important for performance reasons and at the
moment we don't need good performances in RISC-V anyway.

Change-Id: I6a5f7f372e5745a7fcd13e99eb1aca64531ce956