[llcpp] Host LLCPP conformance tests
Host tests are particularly helpful for LLCPP because it is used early
on in the system boot process. When working on benchmarks, I relied on
HLCPP host conformance tests because they used parts of LLCPP (decode),
but it would be useful to have the full thing.

This CL introduces a new optional flag to the fidl build rule
"host_llcpp", which enables building LLCPP on host. This exists because
only some LLCPP features are currently supported on host and building
all FIDL files is not possible.

Test: fx test --host fidl_llcpp_host_conformance_test

Change-Id: I2a049b26a0ff62d608903b6384c55598eb962119
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414097
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>