[testrunner] Record test outputs after each run.
Currently it runs all runs of a test and records all outputs of the
results at the end. However, that causes it to print the "ok"/"not ok"
status of each test run all at the end for a specific test so it's
difficult to see which test logs pertain to which test run.

Multiply: testrunner_tests

Bug: 57814
Change-Id: I03641da41b28253da07e15da74528ee5b7f01c02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414011
Commit-Queue: Ina Huh <ihuh@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gary Miguel <garymm@google.com>
Testability-Review: Oliver Newman <olivernewman@google.com>