[zxdb] SCE #9: Introduce --forward option to continuing
Test: fuchsia-pkg://fuchsia.com/zxc-tests#meta/zxc_unittests.cmx
is a test with death assertions: running it under the debugger will
naturally yield to an exception. `continue` will try to resume the
thread to no effect, while `continue --forward` does indeed forward the
exception to the process-level handlers which resolve it; continuing in
this way, we can reach successfully finish the execution of the test.

Bug: 48767
Change-Id: I5103df87154cee7445cb6fad2b71d69b68115187
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412556
Reviewed-by: Brett Wilson <brettw@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>