[fvm] Remove redundant DEVICE_ADD_INVISIBLE, migrate RemoveDeprecated.
This driver was already migrated to the new DdkInit hook in fxr/407130,
but forgot to remove this arg, which now has no effect.

DdkAsyncRemove is the same as DdkRemoveDeprecated but also calls
the unbind hook. Since the VPartition device implements a trivial
unbind, this has no change in behavior. DdkAsyncRemove does however
assert when trying to remove a device when zxdev_ == nullptr,
so have added a check.

BUG=37831
BUG=43261

Change-Id: I7d0a1b6baa4892875dd41b4467396f1a7a09862b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413354
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Testability-Review: Gianfranco Valentino <gevalentino@google.com>