[zircon][debugger] fix logic in mxcsr_is_valid.
The function `mxcsr_is_valid` had a misleading name: it would return
true if the given value for an mxcsr was invalid, and false if it was
valid.

The code using it used it correctly, a more precise name would have been
`mxcsr_is_invalid`.

This CL modifies the code in mxcsr_is_valid so that it returns true if
the given value for an mxcsr is valid, and false otherwise.

Change-Id: Ic72771475705ce34248f5f69ba44fedb3f88f49a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405487
Commit-Queue: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>
Testability-Review: Brett Wilson <brettw@google.com>