[net-cli] Add route table add/del
This change adds interfaces to net-cli which afford adding and deleting
routes from the route table. This is part of an effort to bring
interfaces that currently only exist in ifconfig into net-cli, with the
aim of eventually deprecating and removing the former.

Fixed: 56758

Change-Id: Ia016c1ba396626018384dc1e1c70c88ecd0217bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414662
Commit-Queue: Amanda Tait <atait@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Testability-Review: Tamir Duberstein <tamird@google.com>