[fx] Remove error message  in fx-env.sh for unknown args
For some WFH workflows, we need to execute scripts with $PATH initialzed
using the $USER's environment. This is done by using the shebang for
those scripts of "#!/bin/bash -i" which creates an interactive shell to
execute the script.

Some developers have added `source fx-env.sh` to their .bashrc files
as we suggest in the getting started guide. This script is parsing
the command line arguments used to start the shell, and incorrectly
assumes that any argurments are from the source command, and not the
command line of bash.

This removes the error message for unknown arguments and they are
ignored.

Change-Id: I78784635b9d2fcd556d40fe42148853cfbfcb6f9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414742
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Testability-Review: Renato Mangini Dias <mangini@google.com>