[blobfs] remove `blob_header_next` from superblock
This field has never been used for ~3 years, and there do not appear to
be any plans to use them. We verify the field is zeroed when mounting
the filesystem.

This change does not modify the value in the field, and continues to
ensure the field is zero on mount, so it is a NOP.

Change-Id: I6cdfd1ead1530baaf7cc3af66b07ec77d5c38c4a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404619
Reviewed-by: Kyo Lee <kyol@google.com>
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Testability-Review: Chris Suter <csuter@google.com>