[wlan] Don't log no-op disconnect_ctx in Inspect
Only log the `disconnect_ctx` if an operation had an effect of moving from
non-idle state to idle state. This is so that the client that consumes
`disconnect_ctx` does not log a disconnect event when it's effectively no-op.

Bug: b/154549756, b/161450584
Test: Logging change only. Manually verified Inspect log.

Change-Id: If0bb4e222f9187744f8ccd4cf6f5c15b2542d496
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412594
Reviewed-by: Marc Khouri <mnck@google.com>
Testability-Review: Marc Khouri <mnck@google.com>
Commit-Queue: Kiet Tran <kiettran@google.com>