[bt][bt-gap] Persist discovery state over active host adapter change
By decoupling the state of Discovery into a dispatcher session and host
session, we are able to provide clients with a dispatcher session while
managing the host sessions at the dispatcher level, allowing us to
seemlessly change the host on which Discovery is ocurring without
perturbing the client's discovery session.

A potential awkwardness with my current implementation is not handling
the acquisition of `HostDiscoverySession` in the `impl DiscoveryState`
methods; as it is, calling `discover_on_active_host` occurs outside the
`DiscoveryState` methods and `HostDiscoverySession` is passed in as an
argument as necessary. It is possible to subsume this functionality
into the `DiscoveryState` methods. This would require passing in a
reference to the `HostDispatcher` to the methods, and potentially
rewriting `discover_on_active_host` to return `impl Future<>` to
avoid borrowing issues.  The advantage to this is that interfacing with
`HostDiscoverySession` will be managed entirely within the
DiscoveryState API.

Bug: 640
Change-Id: I7d1c0736103eba634b5067caa1dc8c985d02ebba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410007
Reviewed-by: Nick Pollard <nickpollard@google.com>
Testability-Review: Nick Pollard <nickpollard@google.com>
Commit-Queue: Ali Saeed <saeedali@google.com>