[audio_core] Fix busy wait in BaseCapturer
If clients do not process packets quickly enough (an underflow),
BaseCapturer::Process gets stuck in a loop where NextMixJob continuously
returns std::nullopt.  This spins and busy waits. To fix this, we now
trigger a condition variable when a packet is inserted into the pending
queue. Process waits on this condition variable when there is no work
immediately available.

Prior to fxr/408098, we shutdown the BaseCapturer when this underflow
was detected. I think that blocking for another packet is nicer
behavior, but I'm open to the idea that we should shutdown instead.

No test added: I'm not sure how to write an automated test for this. I
think the best way to write such a test is to write a framework for
testing audio_core's system load, then setup a lengthy underflow
scenario and verify that audio_core is not burning through CPU. Such a
test would be generally useful for other bugs like fxb/50513. However,
we don't have such a framework yet.

Bug: 57970
Change-Id: I0c2efcc0f3dfb58f4876fe11c65ecb75ea4de021
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/415324
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Testability-Review: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: Tom Bergan <tombergan@google.com>