[sdk] restrict publication of legacy tool paths
The SDK now cross compiles, producing tools in sdk://tools/$arch. For
tools that are cross-compiled they currently appear in
sdk://tools/$tool and sdk://tools/$arch/$tool. This change restricts
the set of tools that appear in sdk://tools/$tool to a known set that
will not grow, enabling the list to be reduced and eventually
removed.

Bug: 42999
Change-Id: I7a9f7e28cdbb8cfee6e31ba037e6a6a97a6487a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414552
Commit-Queue: James Tucker <raggi@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>