[images] support board_tools
Adds support for a |board_tools| arg which can provide a list of tools
to include in the build output.

This is primarily meant for board-specific tooling that may need to be
distributed along with the build files in order to properly use them,
for example a tool to talk to the board in SoC recovery mode for
zero-state flashing or recovery from bricking.

Bug: 56146
Change-Id: I46d69a1b0ba5226943f6bc52a3f18678baaa51f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412700
Reviewed-by: James Tucker <raggi@google.com>
Commit-Queue: David Pursell <dpursell@google.com>