[GN SDK] fserve.sh handles configured device info better.
This change removes the confusion around which device to use with
fserve.

Now, the command line arguments take priority.
if --device-name and --device-ip are not present on the command line,
then the configured device-ip is used if set.
if that is not set, the device-name configured is used.
if that is not set, it uses a device found with device-finder.

Bug: b/163305432
Bug: b/163304430

Change-Id: Ic7532e7907f48cd711c117ab7a8de2d4f3143ec3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/415414
Reviewed-by: Matt Carroll <matthewcarroll@google.com>
Testability-Review: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>