[fuchsia-async] Fix timer bug
Prior to this CL, fuchsia-async timers would occasionally never fire.

This CL fixes a work_scheduler_integration_test flake, but probably also
fixes a bunch of other flakes across the connectivity stack, and
omaha-client and anything that uses the fuchsia-async timer.

Fundamentally, the problem is there was a race in the poll function of
the Timer. When polled, if the timer already fired then the future
should not block. However, if the the timer hasn't yet fired, the
future should block until signalled by the Executor.

The problem is the waker to wake up tasks registered to the timer may be
called between the time it is queried and the time the future's Poll
result is set to Poll::Pending. This may result in the task never
waking up because the Executor drops the timer after triggering it.

The solution is to register the task's waker, then check the did_fire()
condition again to verify whether or not to block.

Bug: 49355
Change-Id: Ib46c3fdaad015c3f2be862e4673e703759e93f71
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/415462
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Fady Samuel <fsamuel@google.com>
Testability-Review: Adam Perry <adamperry@google.com>