[FDF] Exit gracefully when stop event completes
In component lifecycle stop event handler, driver manager suspends all drivers and
If the termination state is Reboot/shutdown, it also issues reboot/shutdown syscall.
At the end of the handler, closing the lifecycle handle causes the component manager
to kill driver manager. When driver manager which is marked critical is killed, the
root job watcher in kernel issues a reboot. If mexec is switched to use orderly shutdown,
this results in an undesired behavior where we want to mexec the system, not reboot.

This changeset exits the driver manager gracefully instead of closing
the lifecycle event handle. If we do so, kernel will not reboot the
system.

Bug: 57448, 52627
Test: fx reboot. fx netboot after moving mexec lib to orderly shutdown.

Change-Id: I5866d0f48c75dfb93d392753b833d45b9e97ef37
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/415478
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Ruchira Ravoori <ravoorir@google.com>