[kernel][clocks] Skip trivial rate adjustment updates.
If a rate adjustment update is requested without a position update,
and the update request is "trivial" (e.g. it is attempting to set the
rate adjustment to what it already is), then skip updating the
transformation equations for a clock.

Fixed: 57593
Change-Id: I63368d2c40a327dea4ea118f16295fc41c60f876
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413390
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>
Commit-Queue: John Grossman <johngro@google.com>
Testability-Review: Martin Puryear <mpuryear@google.com>