[components] Monitor for process exit
Watch the process for exit instead of the channel. The runner
should not watch for the Lifecycle channel to close since this
could cause the runner to kill the component's Job incorrectly.
In particular this prevents us from killing a process which could
be marked critical to the root Job.

Fixed: 56974

TEST=component-manager-tests

Change-Id: I32e45420565a6fe4d5d1084c5af7c9b33aad117b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411877
Commit-Queue: Justin Mattson <jmatt@google.com>
Reviewed-by: Adam Lesinski <adamlesinski@google.com>
Testability-Review: Adam Lesinski <adamlesinski@google.com>