[sshutil] Don't send a keepalive ping on startup
Previously Conn.keepalive() sent a ping as soon as it started, because
the logic to wait for a tick was at the bottom of the for loop, after
the ping was sent. By moving it to the top, we ensure that keepalive()
will only send a ping when the timer fires, which makes testing simpler
and easier to reason about, as well as more explicit.

It's probably not necessary to send an initial ping anyway, since
keepalive() is called immediately after the connection is established.
So there's not much likelihood of losing the connection within the
several milliseconds before that initial keepalive ping is received.
It should be fine to wait for the regular keepalive interval to elapse
before sending the first ping.

I doubt this will fix the flakiness exhibited in http://fxb/50198, but
it will hopefully make it easier to figure out what's causing the
flakiness.

Bug: 50198
Change-Id: I407d4f3f4d62cd908bffa5d42140ffedca71cadd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/415919
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Testability-Review: Oliver Newman <olivernewman@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>